start split
